fix: usar caps PER_USER mas restrictivos y parsear topes formateados
- BenefitDetailPage: cuando hay mas de un cap PER_USER del mismo tipo,
  usar el minimo (mas restrictivo) en vez del primero que aparece en el
  array, para que el orden de los caps no infle el ahorro mostrado ni el
  SavingsSimulator.
- BusinessDetailPage: Number(benefit.tope) daba NaN para topes con
  formato string ('$100.000'). Ahora se parsea el formato argentino y se
  muestra el cap PER_USER monetario solo cuando es el limite mas bajo;
  si no, se conserva el string original del tope (preserva 'Sin tope' y
  el formato).
- Extraidos parseTopeAmount/formatArgentinePeso a src/utils/tope.ts y
  reusados en ambas paginas.
